+++ BsodSimulator/BsodPage.xaml
@@ -1,118 +1,1 @@
-﻿<Page
-    x:Class="BsodSimulator.BsodPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:BsodSimulator"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Page.Resources>
-        <Style TargetType="TextBlock">
-            <Setter Property="Foreground" Value="White" />
-        </Style>
-
-    </Page.Resources>
-    <Grid Background="{x:Bind VM.SelectedColor.Brush, Converter={StaticResource MakeCompilerHappy}, Mode=OneWay}">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*" MaxWidth="120" />
-            <ColumnDefinition Width="4*" />
-            <ColumnDefinition Width="3*" MaxWidth="360" />
-        </Grid.ColumnDefinitions>
-
-        <StackPanel Grid.Column="1">
-            <TextBlock
-                x:Name="textBlock4"
-                Margin="10,15"
-                FontSize="200"
-                Text="{x:Bind VM.Emoji, Mode=OneWay}" />
-            <TextBlock
-                x:Name="textBlock1"
-                Margin="10,15"
-                FontSize="35"
-                Text="{x:Bind VM.Description}"
-                TextWrapping="WrapWholeWords" />
-
-            <TextBlock
-                x:Name="textBlock"
-                Margin="10,15"
-                FontSize="35"
-                Foreground="White"><Run Text="{x:Bind VM.Percentage, Mode=OneWay}" /><Run Text="% Completed" />
-
-            </TextBlock>
-            <StackPanel Margin="15" Orientation="Horizontal">
-
-                <Image
-                    x:Name="image"
-                    Width="200"
-                    Height="200"
-                    HorizontalAlignment="Left"
-                    Source="{x:Bind VM.Bitmap}" />
-
-                <StackPanel Margin="10,-5">
-
-                    <TextBlock x:Name="textBlock2" FontSize="20">
-                        <Run Text="For more information about this issue and possible fixes,visit" />
-                    </TextBlock>
-                    <TextBlock
-                        x:Name="UrlTextBlock"
-                        Margin="0,10"
-                        FontSize="20"
-                        Text="{x:Bind VM.Url}" />
-                    <TextBlock
-                        x:Name="ContactTechnicalTextBlock"
-                        Margin="0,20,0,0"
-                        FontSize="18">
-                        <Run Text="if you call a support person, give them this info:" />
-                    </TextBlock>
-                    <TextBlock
-                        x:Name="textBlock3"
-                        Margin="0,10"
-                        FontSize="18"><Run Text="stop_code:" />
-                        <Run Text="{x:Bind VM.StopCode}" />
-                    </TextBlock>
-                </StackPanel>
-            </StackPanel>
-
-
-        </StackPanel>
-        <VisualStateManager.VisualStateGroups>
-            <VisualStateGroup x:Name="VisualStateGroup">
-                <VisualState x:Name="Wide">
-                    <VisualState.StateTriggers>
-                        <AdaptiveTrigger MinWindowHeight="780" MinWindowWidth="0" />
-
-                    </VisualState.StateTriggers>
-                    <VisualState.Setters>
-                        <Setter Target="textBlock.(TextBlock.FontSize)" Value="35" />
-                        <Setter Target="textBlock1.(TextBlock.FontSize)" Value="35" />
-                        <Setter Target="textBlock2.(TextBlock.FontSize)" Value="20" />
-                        <Setter Target="UrlTextBlock.(TextBlock.FontSize)" Value="20" />
-                        <Setter Target="ContactTechnicalTextBlock.(TextBlock.FontSize)" Value="18" />
-                        <Setter Target="textBlock3.(TextBlock.FontSize)" Value="18" />
-                        <Setter Target="textBlock4.(TextBlock.FontSize)" Value="200" />
-                        <Setter Target="image.(FrameworkElement.Width)" Value="200" />
-                        <Setter Target="image.(FrameworkElement.Height)" Value="200" />
-                    </VisualState.Setters>
-                </VisualState>
-
-                <VisualState x:Name="Narrow">
-                    <VisualState.Setters>
-                        <Setter Target="textBlock.(TextBlock.FontSize)" Value="25" />
-                        <Setter Target="textBlock1.(TextBlock.FontSize)" Value="25" />
-                        <Setter Target="textBlock2.(TextBlock.FontSize)" Value="15" />
-                        <Setter Target="UrlTextBlock.(TextBlock.FontSize)" Value="15" />
-                        <Setter Target="ContactTechnicalTextBlock.(TextBlock.FontSize)" Value="13" />
-                        <Setter Target="textBlock3.(TextBlock.FontSize)" Value="13" />
-                        <Setter Target="textBlock4.(TextBlock.FontSize)" Value="150" />
-                        <Setter Target="image.(FrameworkElement.Width)" Value="150" />
-                        <Setter Target="image.(FrameworkElement.Height)" Value="150" />
-                    </VisualState.Setters>
-                    <VisualState.StateTriggers>
-                        <AdaptiveTrigger MinWindowHeight="0" MinWindowWidth="0" />
-                    </VisualState.StateTriggers>
-                </VisualState>
-            </VisualStateGroup>
-        </VisualStateManager.VisualStateGroups>
-    </Grid>
-</Page>
