+++ BsodSimulator/View/RestartPage.xaml
@@ -1,0 +1,25 @@
+﻿<Page
+    x:Class="BsodSimulator.View.RestartPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:local="using:BsodSimulator.View"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid Background="{x:Bind VM.SelectedColor.Brush}">
+        <StackPanel HorizontalAlignment="Center" VerticalAlignment="Center">
+            <ProgressRing
+                Width="50"
+                Height="50"
+                Foreground="White"
+                IsActive="True" />
+            <TextBlock
+                Margin="20"
+                FontSize="20"
+                Foreground="White">
+                Restarting Windows
+            </TextBlock>
+        </StackPanel>
+    </Grid>
+</Page>
